requirements.txt
turbine-py==1.7.0
scikit-learn==0.24.2
numpy==1.21.2
redis==3.5.3
pydantic==1.8.2
textblob==0.15.3
psycopg2==2.9.1
Faker==8.12.1
hashlib==20081119
requests==2.26.0
influxdb-client==1.20.0
sentry-sdk==1.4.3